requirements.txt
# Core AI + LLM
langchain-openai
openai
chromadb

# Speech-to-text
whisper

# Web app
streamlit

# Data handling
pandas
pydantic
python-dotenv

# Optional utilities
tqdm
requests
